Criação das classes e DTOS relacionadas a pedidos.

+++ src/main/java/com/academia/loja_accenture/modulos/pedido/dto/ProdutoDTO.java
@@ -1,14 +1,13 @@
 package com.academia.loja_accenture.modulos.pedido.dto;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ProdutoDTO (
-    Long id,
-    String nome,
-    String descricao,
-    BigDecimal valor,
-    LocalDateTime createdAt,
-    Long vendedorId
+        Long id,
+        String nome,
+        String descricao,
+        double valor,
+        LocalDateTime createdAt,
+        Long vendedorId
 ) {
 }
